feat(order): order domain - entity
- order status 추가
- order number 추가

+++ order/src/main/java/com/smart/commerce/order/module/order/application/dto/OrderRequest.java
@@ -1,4 +1,5 @@
 package com.smart.commerce.order.module.order.application.dto;
 
-public record OrderRequest(long userId, int storeId) {
+// TODO menu 목록은 cart에 저장된 정보로 가져올지 바로 요청을 받을지
+public record OrderRequest(long userId, long storeId) {
 }
